# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Nine Coupon Server
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
paths:
...
  /coupon/instance:
    x-swagger-router-controller: coupon_instance
    post:
      description: create coupon instance
      operationId: createCouponInstance
      parameters:
        - name: couponTemplateInstance
          in: body
          description: The coupon template instance
          required: true
          schema:
            $ref: '#/definitions/CouponTemplateInstance'
      responses:
        '200':
          description: create coupon instance
          schema:
            $ref: "#/definitions/SuccessJsonResponse"
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/coupon/instance/{appUserId}':
    x-swagger-router-controller: coupon_instance
    get:
      description: get the nine coupons for the user
      operationId: getCouponOfUser
      parameters:
        - name: appUserId
          in: path
          description: The user id
          required: true
          type: string
      responses:
        '200':
          description: Coupon template response
          schema:
            type: array
            items:
              $ref: '#/definitions/CouponInstance'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: get the coupon of lottery
      operationId: receiveTheCoupon
      parameters:
        - name: appUserId
          in: path
          description: The user who receive the coupon
          required: true
          type: string
        - name: couponInstance
          in: body
          description: Which is received
          required: true
          schema:
            $ref: '#/definitions/CouponInstance'
      responses:
        '200':
          description: Coupon is received
          schema:
            $ref: "#/definitions/SuccessJsonResponse"
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  '/coupon/instance/writeoff/{couponInstanceId}':
    x-swagger-router-controller: coupon_instance
    put:
      description: write off the coupon instance
      operationId: writeOffTheCouponInstance
      parameters:
        - name: couponInstanceId
          in: path
          description: The coupon which will be written off
          required: true
          type: string
        - name: couponDetail
          in: body
          schema:
            $ref: '#/definitions/newCouponConsumption'
      responses:
        '200':
          description: Coupon template response
          schema:
            $ref: '#/definitions/CouponConsumption'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/coupon/instance/{clerkUserId}/writeoff':
    x-swagger-router-controller: coupon_instance
    get:
      description: the list of the copon which the user write off 
      operationId: getTheCouponListOfTheClerk
      parameters:
        - name: clerkUserId
          in: path
          required: true
          type: string
      responses:
        "200" :
          description: Coupon template response
          schema:
            type: array
            items:
              $ref: '#/definitions/CouponConsumption'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'          
...
definitions:
...
  Account:
    properties:
      id:
        description: unique identifier for brand
        format: int64
        type: integer
      account:
        type: string
      accountName:
        type: string
      phone:
        type: string
      gender:
        type: integer
      avatar:
        type: string
      createAt:
        type: string
      status:
        description: Account status in the app
        enum:
          - enabled
          - disabled
          - arrears
        type: string
      promoter:
        type: integer
        description: business promoter
      accountType:
        description: Account type in the app
        enum:
          - customer
          - clerk
          - merchant
          - brand
          - market
          - admin
        type: string
    required:
      - id
      - phone
      - accountType
...
  Brand:
    properties:
      id:
        description: unique identifier for brand
        type: integer
        format: int64
      name:
        type: string
      status:
        type: string
        description: Brand status in the app
        enum:
          - enabled
          - disabled
          - arrears
      owner: 
        $ref: "#/definitions/Account"
      createAt:
        type: integer
        format: int64
    required:
      - id
      - name
...
  Location:
    properties:
      latitued:
        type: number
        format: double
      longitued:
        type: number
        format: double
...
  Category:
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
  CouponData:
    properties:
      name:
        type: string
      count:
        type: integer
      brandName:
        type: string
      discount:
        type: number
        format: double
  CouponTemplate:
    properties:
      category:
        $ref: '#/definitions/Category'
      id:
        format: int64
        type: integer
      name:
        type: string
      data:
        $ref: '#/definitions/CouponData'
      status:
        type: string
        enum:
          - enabled
          - disabled
          - auditting
        description: coupon template active status
      origin:
        type: string
      createAt:
        format: int64
        type: integer
    required:
      - id
      - name
      - data
...
  CouponTemplateInstance:
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
      data:
        $ref: '#/definitions/CouponData'
      status:
        type: string
        enum:
          - enabled
          - disabled
          - auditting
        description: coupon template active status
      publishType:
        type: string
      createAt:
        format: int64
        type: integer
      templateInfo:
        $ref: '#/definitions/CouponTemplate'
      brand:
        $ref: '#/definitions/Brand'
    required:
      - id
      - name
      - data
      - templateInfo
      - publishType
      - brand
...
  CouponInstance:
    properties:
      id:
        format: int64
        type: integer
      name:
        type: string
      data:
        $ref: '#/definitions/CouponData'
      status:
        type: string
        enum:
          - enabled
          - disabled
          - auditting
        description: coupon template active status
      templateInstance:
        $ref: '#/definitions/CouponTemplateInstance'
    required:
      - id
      - name
      - data
...
  CouponConsumption:
    properties:
      id:
        description: unique identifier for consumption
        format: int64
        type: integer
      couponInstance:
        $ref: '#/definitions/CouponInstance'
      consumer:
        $ref: '#/definitions/Account'
      clerk:
        $ref: '#/definitions/Account'
      createAt:
        type: integer
        format: int64
      location:
        $ref: '#/definitions/Location'
    required:
      - couponInstance
  newCouponConsumption:
    allOf:
      - $ref: '#/definitions/CouponConsumption'
    required:
      - couponInstance
  Error:
    required:
      - message
      - code
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
  SuccessJsonResponse:
    properties:
      message:
        type: string
      code:
        type: integer
        format: int32
    required:
      - message
...
produces:
  - application/json
  - application/xml
  - text/plain
  - text/html
securityDefinitions:
  oauth2:
    type: oauth2
    scopes:
      read: Read access
      write: Write access
    flow: accessCode
    authorizationUrl: 'http://petstore.swagger.wordnik.com/oauth/authorize'
    tokenUrl: 'http://petstore.swagger.wordnik.com/oauth/token'